requirements.txt
lark-1.2.2
pyparsing-3.2.3
numpy-2.3.0
antlr4
graphviz-0.21
pydot-4.0.1